float -> double, I hate precision problems -_-
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@15511 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzwgen/MSVC/lexer.cxx
@@ -796,12 +796,12 @@ YY_RULE_SETUP
 case 2:
 YY_RULE_SETUP
 #line 29 "..\\..\\src\\lexer.l"
-{ yylval.fl = (float)atof(yytext); return(NUMBER);   }
+{ yylval.fl = (double)atof(yytext); return(NUMBER);   }
 	YY_BREAK
 case 3:
 YY_RULE_SETUP
 #line 30 "..\\..\\src\\lexer.l"
-{ yylval.fl = (float)atof(yytext); return(NUMBER);   }
+{ yylval.fl = (double)atof(yytext); return(NUMBER);   }
 	YY_BREAK
 case 4:
 YY_RULE_SETUP
